Refactor backends into src/backends/ with thin dispatchers
Move backend-specific code into src/backends/{raylib,sokol,none}/.
Merge graphics shared files into single graphics_common.h.
Top-level drawing_helpers.h, font_helper.h, graphics.h are now
simple #ifdef dispatchers.
