--- src/graphics/display_mode.h
@@ -1,9 +1,0 @@
-#pragma once
-
-namespace afterhours {
-namespace graphics {
-
-enum class DisplayMode { Windowed, Headless };
-
-}  // namespace graphics
-}  // namespace afterhours
