+++ src/plugins/e2e_testing/platform_test_input.h
@@ -17,7 +17,7 @@
 #include "input_injector.h"
 #include "visible_text.h"
 
-#include "../../graphics/graphics.h"
+#include "../../graphics.h"
 #include "../../developer.h"
 
 namespace afterhours {
